Changes how files are structured so third party dependencies are only called strictly when needed.
