+++ python/src/adv.py
@@ -1,4 +1,3 @@
-from controllers.main_controller import TerminalController, TkinterController
 from multiprocessing import freeze_support
 from os import terminal_size
 from sys import argv
@@ -11,6 +10,7 @@ if __name__ == "__main__":
   #     we stop our players and exit peacefully
   if len(argv) > 1:
     try:
+        from controllers.main_controller_terminal import TerminalController
         adv = TerminalController()
         adv.main()
     except Exception as e:
@@ -18,5 +18,6 @@ if __name__ == "__main__":
       adv.sound_player.stop_track()
       raise e
   else:
+    from controllers.main_controller_tkinter import TkinterController
     adv = TkinterController()
     adv.start_game()
