+++ python/src/utils/music.py
@@ -1,7 +1,6 @@
 from assets import music_files, sound_files
 from multiprocessing import Process
 import time
-from pygame import mixer
 
 try:
   from utils.playsound_copy import playsound
@@ -135,51 +134,3 @@ class SoundPlayer(Player):
     -   7 -> run away
     """
     return super().play_track(num=num, loop=loop)
-
-
-class PyGamePlayer:
-
-  def __init__(self):
-    mixer.init()
-    self.mixer = mixer
-    self.on = True
-    self.album = []
-
-  def turn_off(self):
-    self.on = False
-
-  def turn_on(self):
-    self.on = True
-
-
-class PyGameMusicPlayer(PyGamePlayer):
-
-  def __init__(self):
-    super().__init__()
-    self.album = music_files
-
-  def play_track(self, num, loop=-1):
-    if not self.on:
-      return
-    track = self.album[num]
-    self.mixer.music.load(track)
-    self.mixer.music.play(loop)
-
-  def stop_track(self):
-    self.mixer.music.stop()
-
-
-class PyGameSoundPlayer(PyGamePlayer):
-
-  def __init__(self):
-    super().__init__()
-    self.album = [self.mixer.Sound(sound) for sound in sound_files]
-
-  def play_track(self, num, loop=-1):
-    if not self.on:
-      return
-    track = self.album[num]
-    self.mixer.Sound.play(track)
-
-  def stop_track(self):
-    pass
